requirements.txt
streamlit==1.28.1
pandas==2.3.3
sqlalchemy==2.0.44
mysql-connector-python==9.5.0
numpy==2.3.4
matplotlib==3.10.7
plotly==5.17.0
pytz==2025.2
python-dateutil==2.9.0.post0
pillow==12.0.0
python-dotenv==1.0.0
